refactor(client): restructure components folder
