+++ packages/client/src/components/Chat/Message.tsx
@@ -1,3 +1,4 @@
+// components
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { AnimatePresence, motion } from 'framer-motion';
