+++ packages/client/src/components/base/ErrorBoundary.tsx
@@ -1,3 +1,4 @@
+// react
 import { Component, ErrorInfo, ReactNode } from 'react';
 
 interface Props {
