+++ packages/client/src/components/base/ErrorFallback.tsx
@@ -1,5 +1,8 @@
+// react
 import { useEffect } from 'react';
 import { useQueryClient } from '@tanstack/react-query';
+
+// components
 import styled from '@emotion/styled';
 
 interface Props {
