+++ packages/client/src/components/home/PostCard.tsx
@@ -1,7 +1,9 @@
 // react
-import styled from '@emotion/styled';
-import { Link } from 'react-router-dom';
 import { PostWithStats } from '~/lib/types';
+import { Link } from 'react-router-dom';
+
+// components
+import styled from '@emotion/styled';
 
 interface Props {
   post: PostWithStats;
