+++ packages/client/src/components/home/skeleton/PostCardSkeleton.tsx
@@ -1,4 +1,4 @@
-// react
+// components
 import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
