+++ packages/client/src/components/post/skeleton/PostContentsSkeleton.tsx
@@ -1,3 +1,4 @@
+// components
 import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
