+++ packages/client/src/pages/RoomPage.tsx
@@ -6,7 +6,7 @@ import { useCreateRoom, useGetRooms } from '~/hooks/queries/room';
 import styled from '@emotion/styled';
 import { Button } from '~/components/common';
 import TabLayout from '~/components/layouts/TabLayout';
-import RoomList from './RoomList';
+import RoomList from '~/components/room/RoomList';
 import AsyncBoundary from '~/components/base/AsyncBoundary';
 import ErrorFallback from '~/components/base/ErrorFallback';
 import { MESSAGE } from '~/constants';
