+++ packages/client/src/pages/SearchPage.tsx
@@ -1,3 +1,4 @@
+// components
 import styled from '@emotion/styled';
 import TabLayout from '~/components/layouts/TabLayout';
